pyproject.toml
[tool.poetry]
name = "binarylane-cli"
version = "0.7.0"
description = ""
authors = ["Nathan O'Sullivan <nathan.osullivan@mammoth.com.au>"]
readme = "README.md"
packages = [{include = "binarylane", from = "src"}, {include = "binarylane", from = "lib" }]

[tool.poetry.dependencies]
python = "^3.8"
terminaltables = "^3.1.10"
httpx = "^0.23.0"
attrs = "^22.2.0"
python-dateutil = "^2.8.2"

[tool.poetry.scripts]
bl = "binarylane.console:app.main"

[tool.poetry.group.dev.dependencies]
openapi-python-client = "^0.11.6"
taskipy = "^1.10.3"
isort = "^5.10.1"
black = "^22.8.0"
pylint = "^2.15.2"
mypy = "^0.971"
pydantic = "^1.10.2"
types-python-dateutil = "^2.8.19"
pytest = "^7.2.0"

[tool.taskipy.tasks]
black = "black ."
isort = "isort ."
mypy = "mypy ."
pylint = "pylint src"
regen = "scripts/generate-client.sh && scripts/generate-commands.sh"
check = "task isort && task black && task mypy && task pylint"
test = "pytest tests"

[tool.isort]
line_length = 120
profile = "black"
known_first_party=["binarylane"]
known_local_folder=["binarylane.console"]

[tool.black]
line-length = 120
target_version = ['py37', 'py38', 'py39']

[tool.pylint.format]
max-line-length = 120
ignore-paths = ['src/binarylane/console/commands/*/']

[tool.pylint.messages_control]
disable = [
    # Most modules contain a single class
    "missing-module-docstring",

    # CLI needs some delayed imports to avoid slow program start
    "import-outside-toplevel",

    # Initial impl still in progress:
    "fixme",
]

[tool.mypy]
plugins = ["pydantic.mypy"]
disallow_any_generics = true
disallow_untyped_defs = true
warn_redundant_casts = true
strict_equality = true
mypy_path = "lib:src"
namespace_packages = true
explicit_package_bases = true
exclude = [
    'src/binarylane/console/actions\.py$',  # from python 3.9
    'src/binarylane/console/commands/.*/.*\.py$',  # FIXME
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
